refactor and name changes
